requirements.txt
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
h11==0.14.0
idna==3.10
isort==5.13.2
mypy-extensions==1.0.0
outcome==1.3.0.post0
psutil==6.1.0
pycparser==2.22
PySocks==1.7.1
python-dotenv==1.0.1
ruff==0.8.2
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
